+++ Libraries/KinCharacteristicsTypes.h
@@ -7,13 +7,13 @@ NS_CORE_START
 
 struct KpdValue
 {
-	float _kpdA;
-	float _kpdB;
+	double _kpdA;
+	double _kpdB;
 };
 
 typedef std::map<NS_CORE Element, int>		Z;
-typedef std::map<NS_CORE Element, float>	M;
-typedef std::map<NS_CORE Element, float>	W;
+typedef std::map<NS_CORE Element, double>	M;
+typedef std::map<NS_CORE Element, double>	W;
 typedef std::map<NS_CORE Element, char>		N;
 typedef std::map<NS_CORE GearSetNumber, KpdValue> KpdZac;
