+++ Libraries/Singletons.h
@@ -59,7 +59,7 @@ public:
 
 	void										setGlobalParameters( const size_t w, const size_t n, const size_t d );
 	void										addRangeK( const Range& range );
-	void										addGearRatio( const float& i );
+	void										addGearRatio( const double& i );
 
 };
