+++ SintezPPKinCharacteristics/Function.h
@@ -20,10 +20,10 @@ private:
 	int											m_na;
 	int											m_nb;
 	int											m_nh;
-	float										m_a;
-	float										m_b;
-	float										m_c;
-	float										m_kpdSum;
+	double										m_a;
+	double										m_b;
+	double										m_c;
+	double										m_kpdSum;
 
 	Function() = delete;
 	Function( const NS_CORE N& n, const NS_CORE GearSetNumber gearSet );
